backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
geoalchemy2==0.14.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.27.0
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.2.0
APScheduler==3.10.4
